+++ src/Black/Component/Page/Infrastructure/CQRS/Handler/PublishWebPageHandler.php
@@ -10,10 +10,11 @@
 
 namespace Black\Component\Page\Infrastructure\CQRS\Handler;
 
+use Black\Component\Page\Domain\Event\WebPageDepublishedEvent;
 use Black\Component\Page\Infrastructure\CQRS\Command\PublishWebPageCommand;
 use Black\Component\Page\Infrastructure\Doctrine\WebPageManager;
-use Black\Component\Page\Infrastructure\DomainEvent\WebPagePublishedEvent;
-use Black\Component\Page\Infrastructure\DomainEvent\WebPagePublishedSubscriber;
+use Black\Component\Page\Domain\Event\WebPagePublishedEvent;
+use Black\Component\Page\Infrastructure\Listener\WebPagePublishedListener;
 use Black\Component\Page\Infrastructure\Service\WebPageWriteService;
 use Black\DDD\CQRSinPHP\Infrastructure\CQRS\CommandHandler;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -42,7 +43,7 @@ final class PublishWebPageHandler implements CommandHandler
     protected $eventDispatcher;
 
     /**
-     * @var \Black\Component\Page\Infrastructure\DomainEvent\WebPagePublishedSubscriber
+     * @var \Black\Component\Page\Infrastructure\Listener\WebPagePublishedListener
      */
     protected $subscriber;
 
@@ -50,18 +51,18 @@ final class PublishWebPageHandler implements CommandHandler
      * @param WebPageWriteService $service
      * @param WebPageManager $manager
      * @param EventDispatcherInterface $eventDispatcher
-     * @param WebPagePublishedSubscriber $subscriber
+     * @param WebPagePublishedListener $listener
      */
     public function __construct(
         WebPageWriteService $service,
         WebPageManager $manager,
         EventDispatcherInterface $eventDispatcher,
-        WebPagePublishedSubscriber $subscriber
+        WebPagePublishedListener $listener
     ) {
         $this->service         = $service;
         $this->manager         = $manager;
         $this->eventDispatcher = $eventDispatcher;
-        $this->subscriber      = $subscriber;
+        $this->listener      = $listener;
     }
 
     /**
@@ -75,7 +76,7 @@ final class PublishWebPageHandler implements CommandHandler
         $this->manager->flush();
 
         $event = new WebPagePublishedEvent($page->getWebPageId()->getValue(), $page->getName());
-        $this->eventDispatcher->addSubscriber($this->subscriber);
+        $this->eventDispatcher->addSubscriber($this->listener);
         $this->eventDispatcher->dispatch('web_page.published', $event);
     }
 }
