+++ src/Black/Component/Page/Infrastructure/Listener/WebPageDepublishedListener.php
@@ -8,18 +8,19 @@
  * file that was distributed with this source code.
  */
 
-namespace Black\Component\Page\Infrastructure\DomainEvent;
+namespace Black\Component\Page\Infrastructure\Listener;
 
+use Black\Component\Page\Domain\Event\WebPageDepublishedEvent;
 use Monolog\Logger;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
- * Class WebPageDepublishedSubscriber
+ * Class WebPageDepublishedListener
  *
  * @author  Alexandre 'pocky' Balmes <alexandre@lablackroom.com>
  * @license http://opensource.org/licenses/mit-license.php MIT
  */
-class WebPageDepublishedSubscriber implements EventSubscriberInterface
+class WebPageDepublishedListener implements EventSubscriberInterface
 {
     /**
      * @param Logger $logger
